Dump then load notice output to convert Numpy to native types
The check `assert actual == expected` will fail if the `actual`
data structure contains Numpy arrays. The exception reads,
`ValueError: The truth value of an array with more than one element
is ambiguous. Use a.any() or a.all()`.

Dump the notice to JSON and then load it again to convert all
Numpy arrays to native data types so that we can rely on Python's
built-in deep equality check for dicts and lists.